feat(murmur3): improve with better inline fns

+++ tests/test_strukts_hashing.cpp
@@ -4,17 +4,18 @@
 
 #include "gtest/gtest.h"
 #include "strukts_hashing.h"
+#include "strukts_types.h"
 
 namespace {
     TEST(STRUKTS_HASHING_SUITE, SHOULD_32BIT_MURMUR3_HASH_STRING_WITH_5_BYTES) {
         // arrange
         const char str[10] = "abcde";  // literal str: {'a', 'b', 'c', 'd', 'e', '\0'};
-        const uint8_t* str_bytes = (uint8_t*)str;
-        const uint32_t seed = 0;
+        const BYTE* str_bytes = (BYTE*)str;
+        const WORD seed = 0;
         size_t str_len = strlen(str);
 
         // act
-        uint32_t hash = strukts_murmur3_hash(str_bytes, str_len, seed);
+        WORD hash = strukts_murmur3_hash(str_bytes, str_len, seed);
 
         // assert
         EXPECT_EQ(hash, -392455434);  // expectation from python's lib mmh3
@@ -23,12 +24,12 @@ namespace {
     TEST(STRUKTS_HASHING_SUITE, SHOULD_32BIT_MURMUR3_HASH_STRING_WITH_2_BYTES) {
         // arrange
         const char str[10] = "ab";
-        const uint8_t* str_bytes = (uint8_t*)str;
-        const uint32_t seed = 5;
+        const BYTE* str_bytes = (BYTE*)str;
+        const WORD seed = 5;
         size_t str_len = strlen(str);
 
         // act
-        uint32_t hash = strukts_murmur3_hash(str_bytes, str_len, seed);
+        WORD hash = strukts_murmur3_hash(str_bytes, str_len, seed);
 
         // assert
         EXPECT_EQ(hash, 810406479);  // expectation from python's lib mmh3
@@ -37,12 +38,12 @@ namespace {
     TEST(STRUKTS_HASHING_SUITE, SHOULD_32BIT_MURMUR3_HASH_STRING_WITH_22_BYTES) {
         // arrange
         const char str[50] = "some really nice long key";
-        const uint8_t* str_bytes = (uint8_t*)str;
-        const uint32_t seed = 10;
+        const BYTE* str_bytes = (BYTE*)str;
+        const WORD seed = 10;
         size_t str_len = strlen(str);
 
         // act
-        uint32_t hash = strukts_murmur3_hash(str_bytes, str_len, seed);
+        WORD hash = strukts_murmur3_hash(str_bytes, str_len, seed);
 
         // assert
         EXPECT_EQ(hash, -464589223);  // expectation from python's lib mmh3
